+++ app/Models/Employee.php
@@ -370,266 +370,4 @@ class Employee extends Model
     {
         return anchorNewTab(employeeInListsLinkUrl($this->id), $this->name);
     }
-
-    /**
-     * @param  orderBy: asc / desc
-     * @return collection
-     */
-    // TODO:: wip, remove this method, bec. this was transfered in DTR service class
-    public function logs($date = null, $logTypes = null, $orderBy = 'asc') 
-    {
-        $logs = null;
-        $date = ($date == null) ? currentDate() : $date;
-        $shiftToday = $this->shiftDetails($date); 
-        
-        if ($logTypes == null) {
-            $logTypes = dtrLogTypes();
-        }else { 
-            if (!is_array($logTypes)) {
-                $logTypes = (array) $logTypes;
-            }
-        }
-
-        if ($shiftToday) {
-            if (!$shiftToday->open_time) {
-                // !open_time
-                $logs = $this->dtrLogs()
-                    ->whereBetween('log', [$shiftToday->relative_day_start, $shiftToday->relative_day_end])
-                    ->whereIn('dtr_log_type_id', $logTypes);
-            }else {
-                // open_time
-                $logs = $this->dtrLogs()
-                    ->whereDate('log', '=', $shiftToday->date)
-                    ->whereIn('dtr_log_type_id', $logTypes);
-
-                //deduct 1 day to date and if not open_time, be sure to add whereNotBetween to avoid retrieving prev. logs.
-                $prevShift = $this->shiftDetails(subDaysToDate($shiftToday->date));
-                if ($prevShift && !$prevShift->open_time) {
-                    $logs = $logs->whereNotBetween('log', [$prevShift->relative_day_start, $prevShift->relative_day_end]);
-                }
-
-                // return compact('prevShift', 'shiftToday', 'logs'); // NOTE:: for debug only
-            }
-        }
-
-        if ($logs) {
-            return $logs->orderBy('log', $orderBy)->get();
-        }
-
-        return $logs;
-    }
-
-    // TODO:: wip, remove this method, bec. this was transfered in DTR service class
-    public function shiftDetails($date)
-    {
-        $shiftDetails = null;
-
-        $shift = $this->employeeShiftSchedules()->date($date)->first();
-        if ($shift) {
-            $shiftDetails = $shift->details($date);
-        }
-
-        $changeShift = $this->changeShiftSchedules()->date($date)->first();
-        if ($changeShift) {
-            // if todays date has employee changeshift then return that instead
-            $shiftDetails = $changeShift->shiftSchedule()->first();
-        }
-        
-        if ($shiftDetails) {
-            $shiftDetails->date = $date;
-            $dbRelativeDayStart = $shiftDetails->relative_day_start;
-            unset($shiftDetails->relative_day_start); // i unset this obj. property and added again at the bottom to chnage order.
-            $shiftDetails->db_relative_day_start = $dbRelativeDayStart; 
-            $shiftDetails->start_working_at = null;
-            $shiftDetails->end_working_at = null; // custom object
-            $shiftDetails->relative_day_start = null;
-            $shiftDetails->relative_day_end = null;
-
-            if (!$shiftDetails->open_time) {
-
-                // custom/added obj properties
-                $shiftDetails->start_working_at = $date .' '.$shiftDetails->start_working_hours;
-                $shiftDetails->end_working_at = $date .' '.$shiftDetails->end_working_hours;
-                
-                if (carbonInstance($shiftDetails->end_working_at)->lessThan($shiftDetails->start_working_at)) {
-                    $shiftDetails->end_working_at = addDaysToDate($date) .' '.$shiftDetails->end_working_hours;
-                }
-
-                $shiftDetails->relative_day_start = $date . ' '.$dbRelativeDayStart;
-
-                if (carbonInstance($shiftDetails->relative_day_start)->greaterThan($date.' '.$shiftDetails->start_working_at)) {
-                    $shiftDetails->relative_day_start = subDaysToDate($date). ' '.$dbRelativeDayStart;
-                }
-                $shiftDetails->relative_day_end = carbonInstance($shiftDetails->relative_day_start)->addDay()->format('Y-m-d H:i');
-            }else {
-                // over shift is open time set WH and OH to null
-                $shiftDetails->working_hours = null;
-                $shiftDetails->overtime_hours = null;
-            }
-
-
-            // working_hours_with_date init
-            $shiftDetails->working_hours_with_date = null;
-
-            if ($shiftDetails->working_hours) {
-                $shiftDetails->working_hours = $shiftDetails->working_hours['working_hours'];
-                
-                // assign value to working_hours_with_date
-                $shiftDetails->working_hours_with_date = collect($shiftDetails->working_hours)
-                    ->mapWithKeys(function ($item, $key) use ($date) {
-            
-                        $whStart =  $date .' '.$item['start'];
-                        $whEnd =  $date .' '.$item['end'];
-                        
-                        if (carbonInstance($whEnd)->lessThan($whStart)) {
-                            $whEnd = addDaysToDate($date) .' '.$item['end'];
-                        }
-
-                        return [
-                            $key => [
-                                'start' => $whStart,
-                                'end' => $whEnd,
-                            ]
-                        ];
-                    })->toArray();
-                // end assign value to working_hours_with_date
-            }
-
-            if ($shiftDetails->overtime_hours) {
-                $shiftDetails->overtime_hours = $shiftDetails->overtime_hours['overtime_hours'];
-            }
-
-
-            $detailsText = "";
-            $detailsText .= "Name : $shiftDetails->name\n";
-            $detailsText .= "Open Time : ".booleanOptions()[$shiftDetails->open_time]."\n";
-
-            $detailsText .= "Working Hours :\n";
-            if (count($shiftDetails->working_hours_in_array) > 0) {
-                $temp = "   ".implode(",\n   ", $shiftDetails->working_hours_in_array);
-                $detailsText .= $temp."\n";
-            }
-
-            $detailsText .= "Overtime Hours :\n";
-            if (count($shiftDetails->overtime_hours_in_array) > 0) {
-                $temp = "   ".implode(",\n   ", $shiftDetails->overtime_hours_in_array);
-                $detailsText .= $temp."\n";
-            }
-
-
-            $detailsText .= "Dynamic Break : ".booleanOptions()[$shiftDetails->dynamic_break]."\n"; 
-            $detailsText .= "Dynamic Break Credit : $shiftDetails->dynamic_break_credit\n";
-
-
-            $detailsText .= "Relative Day Start : ".carbonDateTimeFormat($shiftDetails->relative_day_start)."\n";
-            $detailsText .= "Relative Day End : ".carbonDateTimeFormat($shiftDetails->relative_day_end)."\n";
-            
-            $shiftDetails->details_text = $detailsText;
-        }// end if $shiftDetails
-
-        return $shiftDetails;   
-    }
-
-    // TODO:: wip, 
-    public function shiftToday()
-    {
-        $date = currentDate();
-        $currentShift = $this->shiftDetails($date);
-        $prevShift = $this->shiftDetails(subDaysToDate($date, 1));
-        $currentDateTime = currentDateTime();
-
-        //return compact('currentDateTime', 'currentShift', 'prevShift'); // NOTE:: comment this, for debug only
-
-        // currentShift not open_time
-        if ($currentShift && !$currentShift->open_time) {
-            $dayStart = $currentShift->relative_day_start;
-            $dayEnd = $currentShift->relative_day_end;
-            if (carbonInstance($currentDateTime)->betweenIncluded($dayStart, $dayEnd)) {
-                return $currentShift;
-            }
-        }
-
-        // prevShift not open_time
-        if ($prevShift && !$prevShift->open_time) {
-            $dayStart = $prevShift->relative_day_start;
-            $dayEnd = $prevShift->relative_day_end;
-            if (carbonInstance($currentDateTime)->betweenIncluded($dayStart, $dayEnd)) {
-                return $prevShift;
-            }
-        }
-
-        if ($currentShift) {
-            // currentShift open_time
-            if ($currentShift->open_time) {
-                return $currentShift;
-            }
-
-            // prevShift open_time
-            if ($prevShift && $prevShift->open_time) {
-                return $prevShift;
-            }
-        }
-
-        return;
-    }
-
-    /**
-     * show or hide Employee Time Clock buttons.
-     * @return associative array
-     */
-    // TODO:: wip, 
-    public function timeClock()
-    {
-        $in         = false;
-        $out        = false;
-        $breakStart = false;
-        $breakEnd   = false;
-        $hasShift   = false;
-
-        $shiftToday = $this->shiftToday();
-        
-        if ($shiftToday) {
-            $logsToday = $this->logs($shiftToday->date, [1,2]); // 1 = in, 2 = OUT
-            $breaksToday = $this->logs($shiftToday->date, [3,4]); // 3 = break start , 4 = break end 
-            $hasShift = true;
-
-            // in
-            if (($logsToday->last() == null) || $logsToday->last()->dtr_log_type_id == 2) {
-                $in = true;
-            }
-
-            // out
-            if ($logsToday->last() && $logsToday->last()->dtr_log_type_id == 1) {
-                $out = true;
-            }
-            
-            // break start
-            if ($out && $shiftToday->dynamic_break && $breaksToday->last() == null) {
-                $breakStart = true;
-            }
-            
-            // break end
-            if ($out && $breaksToday->last() && $breaksToday->last()->dtr_log_type_id == 3) {    
-                $breakEnd = true;
-                $out = false;
-            }
-           
-            // logs in/out limit
-            $outLimit = count($shiftToday->working_hours);
-            $logOuts = $this->logs($shiftToday->date, [2]); // 2 = Out
-            $totalOutLogs = ($logOuts != null) ? count($logOuts) : 0; 
-            if ($totalOutLogs >= $outLimit) {
-                $in = false;
-            }
-        }
-
-        return [
-            'hasShift'   => $hasShift,
-            'in'         => $in,
-            'out'        => $out,
-            'breakStart' => $breakStart,
-            'breakEnd'   => $breakEnd,
-        ];
-    }
-
 }
