+++ app/Services/Traits/LogTrait.php
@@ -10,11 +10,14 @@ trait LogTrait
      * @return collection
      * * NOTE:: Require shiftDetails and employee property instance
      */
-    public function logs($date = null, $logTypes = null, $orderBy = 'asc') 
+    public function logs($logTypes = null, $orderBy = 'asc') 
     {
+        // if no shift
+        if (!$this->shiftDetails) {
+            return;
+        }
+        
         $logs = null;
-        $date = ($date == null) ? currentDate() : $date;
-        $shiftToday = $this->shiftDetails;
         
         if ($logTypes == null) {
             $logTypes = dtrLogTypes();
@@ -24,28 +27,27 @@ trait LogTrait
             }
         }
 
-        if ($shiftToday) {
-            if (!$shiftToday->open_time) {
-                // not open_time
-                $logs = $this->employee->dtrLogs()
-                    ->with('dtrLogType')
-                    ->whereBetween('log', [$shiftToday->relative_day_start, $shiftToday->relative_day_end])
-                    ->whereIn('dtr_log_type_id', $logTypes);
-            }else {
-                // open_time
-                $logs = $this->employee->dtrLogs()
-                    ->with('dtrLogType')
-                    ->whereDate('log', '=', $shiftToday->date)
-                    ->whereIn('dtr_log_type_id', $logTypes);
-
-                //deduct 1 day to date and if not open_time, be sure to add whereNotBetween to avoid retrieving prev. logs.
-                $prevShift = $this->shiftDetails(subDaysToDate($shiftToday->date));
-                if ($prevShift && !$prevShift->open_time) {
-                    $logs = $logs->whereNotBetween('log', [$prevShift->relative_day_start, $prevShift->relative_day_end]);
-                }
-
-                // return compact('prevShift', 'shiftToday', 'logs'); // NOTE:: for debug only
+        if (!$this->shiftDetails->open_time) {
+            // not open_time
+            $logs = $this->employee->dtrLogs()
+                ->with('dtrLogType')
+                ->whereBetween('log', [$this->shiftDetails->relative_day_start, $this->shiftDetails->relative_day_end])
+                ->whereIn('dtr_log_type_id', $logTypes);
+        }else {
+            // open_time
+            $logs = $this->employee->dtrLogs()
+            ->with('dtrLogType')
+            ->whereDate('log', '=', $this->shiftDetails->date)
+            ->whereIn('dtr_log_type_id', $logTypes);
+            
+            //deduct 1 day to date and if not open_time, be sure to add whereNotBetween to avoid retrieving prev. logs.
+            // TODO:: wip, test on open_time shift
+            $prevShift = $this->shiftDetails(subDaysToDate($this->shiftDetails->date));
+            if ($prevShift && !$prevShift->open_time) {
+                $logs = $logs->whereNotBetween('log', [$prevShift->relative_day_start, $prevShift->relative_day_end]);
             }
+
+            // return compact('prevShift', 'this->shiftDetails', 'logs'); // NOTE:: for debug only
         }
 
         if ($logs) {
